refactor(logic): CarRacingCore 리팩토링
모델 변경에 따라 구조 변경
