+++ src/main/java/com/rick/racing/core/CarRacingCore.java
@@ -13,24 +13,23 @@ public class CarRacingCore {
     public static final int RANDOM_BOUND = 10;
     public static final int GO_THRESHOLD = 4;
     private final InputView inputView;
-    private final ResultView resultView;
     private final Random random;
 
     public CarRacingCore() {
         inputView = new InputView();
-        resultView = new ResultView();
         random = new Random();
     }
 
     public void start() {
         RacingPlayData racingPlayData = inputView.getGameDataFromUser();
-        resultView.showResult(racingPlayData, doGame(racingPlayData, this::isGo));
+        RacingResult racingResult = doGame(racingPlayData, this::isGo);
+        new ResultView(racingPlayData, racingResult).showResult();
     }
 
     public RacingResult doGame(RacingPlayData racingPlayData, Supplier<Boolean> goOrStopSupplier) {
         RacingResult racingResult = new RacingResult();
 
-        for (int carIndex = 0; carIndex < racingPlayData.getCarCount(); carIndex++) {
+        for (int carIndex = 0; carIndex < racingPlayData.carCount(); carIndex++) {
             CarRecordHistory carRecordHistory = playRacing(racingPlayData, goOrStopSupplier);
             racingResult.addRecord(carRecordHistory);
         }
@@ -38,14 +37,23 @@ public class CarRacingCore {
         return racingResult;
     }
 
-    private CarRecordHistory playRacing(RacingPlayData racingPlayData,  Supplier<Boolean> goOrStopSupplier) {
+    private CarRecordHistory playRacing(RacingPlayData racingPlayData, Supplier<Boolean> goOrStopSupplier) {
         CarRecordHistory carRecordHistory = new CarRecordHistory();
 
-        for (int tryIndex = 0; tryIndex < racingPlayData.getTryCount(); tryIndex++) {
-            carRecordHistory.addHistory(goOrStopSupplier.get());
+        for (int tryIndex = 0; tryIndex < racingPlayData.tryCount(); tryIndex++) {
+            move(carRecordHistory, goOrStopSupplier);
         }
 
         return carRecordHistory;
+    }
+
+    private void move(CarRecordHistory carRecordHistory, Supplier<Boolean> goOrStopSupplier) {
+        if(goOrStopSupplier.get()) {
+            carRecordHistory.go();
+            return;
+        }
+
+        carRecordHistory.stop();
     }
 
     private boolean isGo() {
