+++ src/test/java/com/rick/racing/core/CarRacingCoreTest.java
@@ -1,9 +1,7 @@
-package com.rick.racing;
+package com.rick.racing.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.rick.racing.core.CarRacingCore;
-import com.rick.racing.model.CarRecordHistory;
 import com.rick.racing.model.RacingPlayData;
 import com.rick.racing.model.RacingResult;
 import java.util.List;
@@ -11,7 +9,7 @@ import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class CarRacingTest {
+public class CarRacingCoreTest {
 
     private final CarRacingCore carRacing = new CarRacingCore();
 
@@ -38,11 +36,16 @@ public class CarRacingTest {
         RacingResult racingResult = carRacing.doGame(racingPlayData, this::isGo);
 
         for (int carIndex = 0; carIndex < carCount; carIndex++) {
-            final CarRecordHistory carRecordHistory = racingResult.getRecord(carIndex);
-            for (int tryIndex = 0; tryIndex < tryCount; tryIndex++) {
-                assertThat(carRecordHistory.getPosition(tryIndex)).isEqualTo(
-                    expected.get(carIndex).get(tryIndex));
-            }
+            checkCarRecordHistories(tryCount, carIndex, expected, racingResult);
+        }
+    }
+
+    private void checkCarRecordHistories(int tryCount, int carIndex, List<List<Integer>> expected, RacingResult racingResult) {
+        for (int tryIndex = 0; tryIndex < tryCount; tryIndex++) {
+            assertThat(racingResult.recordPosition(carIndex, tryIndex))
+                .isEqualTo(
+                    expected.get(carIndex)
+                        .get(tryIndex));
         }
     }
